refactor(ui-effects): move the 3 product effects into @/shared/ui/effects
