+++ src/shared/ui/effects/index.ts
@@ -1,0 +1,3 @@
+export { MatrixRain, MatrixRainOverlay } from "./matrix-rain";
+export { MatrixText } from "./matrix-text";
+export { GlitchText } from "./glitch-text";
