+++ src/shared/ui/fold-block.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import { useState, type ReactNode } from "react";
-import { MatrixText } from "./matrix-text";
+import { MatrixText } from "@/shared/ui/effects";
 import { prefersReducedMotion } from "@/shared/lib/prefers-reduced-motion";
 
 interface FoldBlockProps {
