+++ src/shared/ui/index.ts
@@ -5,7 +5,6 @@ export { StatusBadge } from "./status-badge";
 export type { Status } from "./status-badge";
 export { BrokenHeartIcon } from "./heart-icons";
 export { Dot } from "./dot";
-export { MatrixText } from "./matrix-text";
 export { Menu } from "./menu";
 export type { MenuItem } from "./menu";
 export { Field } from "./field";
@@ -22,7 +21,6 @@ export { IconSubmitButton } from "./icon-submit-button";
 export { Loading, Spinner } from "./loading";
 export { ProgressBar } from "./progress-bar";
 export { StatBar } from "./stat-bar";
-export { GlitchText } from "./glitch-text";
 export { Stepper } from "./stepper";
 export { TabNav } from "./tab-nav";
 export type { TabItem } from "./tab-nav";
@@ -30,7 +28,6 @@ export { UnderlineTabs } from "./underline-tabs";
 export type { UnderlineTabItem } from "./underline-tabs";
 export { Console, ConsoleTitleBar } from "./console";
 export { InfoBox } from "./info-box";
-export { MatrixRain, MatrixRainOverlay } from "./matrix-rain";
 export { RevealMark } from "./reveal-mark";
 export { Kbd } from "./kbd";
 export { ScanText } from "./scan-text";
