+++ src/shared/ui/post-body.tsx
@@ -2,8 +2,7 @@ import { Children, type ComponentPropsWithoutRef, type ReactNode } from "react";
 import ReactMarkdown, { type Components } from "react-markdown";
 import remarkGfm from "remark-gfm";
 import remarkDirective from "remark-directive";
-import { GlitchText } from "./glitch-text";
-import { MatrixText } from "./matrix-text";
+import { GlitchText, MatrixText } from "@/shared/ui/effects";
 import { RevealMark } from "./reveal-mark";
 import { Kbd } from "./kbd";
 import { ScanText } from "./scan-text";
